rentalservices-backend-develop/backend/requirements.txt
alembic==1.9.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==1.4.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.4
twilio==6.50.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
Flask-Mail==0.9.1
pytz==2022.7.1
